+++ packages/amplify-category-function/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@4.1.1...amplify-category-function@4.1.2) (2022-10-03)
+
+**Note:** Version bump only for package amplify-category-function
+
+
+
+
+
 ## [4.1.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@4.0.12...amplify-category-function@4.1.1) (2022-09-21)
 
 **Note:** Version bump only for package amplify-category-function